src/app/shared/custom-input-controls/file-uploader/file-uploader.component.spec.ts
// import {
//   beforeEach,
//   beforeEachProviders,
//   describe,
//   expect,
//   it,
//   inject,
// } from '@angular/core/testing';
// import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
// import { Component } from '@angular/core';
// import { By } from '@angular/platform-browser';
// import { FileUploaderComponent } from './file-uploader.component';

// describe('Component: FileUploader', () => {
//   let builder: TestComponentBuilder;

//   beforeEachProviders(() => [FileUploaderComponent]);
//   beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
//     builder = tcb;
//   }));

//   it('should inject the component', inject([FileUploaderComponent],
//       (component: FileUploaderComponent) => {
//     expect(component).toBeTruthy();
//   }));

//   it('should create the component', inject([], () => {
//     return builder.createAsync(FileUploaderComponentTestController)
//       .then((fixture: ComponentFixture<any>) => {
//         let query = fixture.debugElement.query(By.directive(FileUploaderComponent));
//         expect(query).toBeTruthy();
//         expect(query.componentInstance).toBeTruthy();
//       });
//   }));
// });

// @Component({
//   selector: 'test',
//   template: `
//     <app-file-uploader></app-file-uploader>
//   `,
//   directives: [FileUploaderComponent]
// })
// class FileUploaderComponentTestController {
// }

